Add working JavaFX prototype, expecting refactoring and reorganisation
