+++ text-ui-test/runtest.bat
@@ -17,7 +17,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin catbot.CatBot < input.txt > ACTUAL.TXT
+java -classpath ..\bin catbot.bot.CatBot < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
